Changed SimulatedDevice to be loaded as a plugin through pluginlib. Chnaged SimDev_Echo accordingly.
